backend/shared/intjtech-logging/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "intjtech-logging"
version = "0.1.0"
description = "Unified logging library for INTJTech services"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-json-logger>=2.0.0",
]

[project.optional-dependencies]
fastapi = [
    "fastapi>=0.100.0",
    "starlette>=0.27.0",
]
loki = [
    "python-logging-loki>=0.3.1",
]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
]

[tool.setuptools.packages.find]
where = ["."]
